requirements.txt
pdfplumber
pandas
google-genai
python-dotenv
openpyxl
Pillow
